refactor(validation): extract AJV error postprocessing
